INGESTION/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.38.37",
    "opensearch-py>=3.0.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
]
